backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "fastapi>=0.135.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "uvicorn>=0.42.0",
]
